Added notes, fixed bugs and created new header for stuff needed for visual representation of possible steps
