+++ CMakeLists.txt
@@ -25,6 +25,7 @@ add_executable(
     cell.cpp 
     board.cpp 
     figure.cpp
+    steps_representation.cpp
     ${LIB_DIR}/Graph_lib/Graph.cpp
     ${LIB_DIR}/Graph_lib/GUI.cpp
     ${LIB_DIR}/Graph_lib/Window.cpp
