+++ cell.h
@@ -9,10 +9,11 @@ constexpr int DFTBOF = 30; //distance_from_the_beginning_of_coordinates
 
 struct Checker;
 
+//stores coordinates on a chessboard from ('a', 1) to ('h', 8)
 struct Coordinate
 {
-    Coordinate(int x_, int y_): x{x_}, y{y_} {}
-    int x;
+    Coordinate(char x_, int y_): x{x_}, y{y_} {}
+    char x;
     int y;
 };
 
@@ -51,11 +52,10 @@ struct Cell : Graph_lib::Button
 
     bool has_figure () const { return figure != nullptr; }
 
+    //removed const and i don't know how it will turn out
     /*const*/ Figure& get_figure () const;  // обязательно нужна проверка не нулевой ли указатель checker
-                                            // removed const and i don't know how it will turn out
 
-    Coordinate location(); // получаем пару {x, y} - координату положения клетки
-
+    Coordinate location(); //gets pair {char x , int y} - coordinate of a cell
   private:
     Type type;
     void reset_color ();
